// src/types.ts
export interface IDatas {
   map(
      arg0: (typedData: any) => import('react/jsx-runtime').JSX.Element
   ): import('react').ReactNode;
   collectionId: string;
   collectionName: string;
   content: string;
   created: string;
   id: string;
   img: string[];
   title: string;
   updated: string;
}